Add changes to binary dump exporter
